% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomials1.r
\name{binomials1}
\alias{binomials1}
\title{Demonstration of binomial distributions}
\usage{
binomials1(n = 5, p = 0.5, x.axis.type = "Gaussian",
  y.axis.type = "Full Range", plot.type = "Counts",
  add.normal = FALSE, add.values = FALSE, crit.value.c = 1,
  crit.value.p = 0.5, show.critical = FALSE)
}
\arguments{
\item{n}{Integer for number of flips}

\item{p}{Probability of success (a value between 0 and 1)}

\item{x.axis.type}{("Gaussian","Full Range")}

\item{y.axis.type}{("Full Range","Fixed")}

\item{plot.type}{("Counts","Proportions")}

\item{add.normal}{Logical (False)}

\item{add.values}{Logical (False)}

\item{crit.value.c}{Critical value for counts}

\item{crit.value.p}{Critical value for probability}

\item{show.critical}{Logical (False)}
}
\value{
None
}
\description{
Demonstration of binomial distributions
}
\examples{
binomials()

}
