% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_hist_source.r
\name{demo_hist}
\alias{demo_hist}
\title{Illustration of histogram}
\usage{
demo_hist(x = sunspots)
}
\arguments{
\item{x}{numeric data}
}
\value{
None
}
\description{
Creates a histogram of using default R settings.  This function is used only to test
package installation.
}
\examples{
demo_hist(sunspots)

}
