% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_nhst_source.r
\name{demo_nhst}
\alias{demo_nhst}
\title{Demonstration of 2-sample inference}
\usage{
demo_nhst(demo.type = 1, nreps = 100, sleep.sec = 0.3,
  add.diff = 3.75, show.ttest = F)
}
\arguments{
\item{demo.type}{a number (1=show actual groups 2=run random shuffling)}

\item{nreps}{a number (10-100) that controls number of replications}

\item{sleep.sec}{a number (0-3) that controls waiting between replications}

\item{add.diff}{number specifying the mean difference the two groups}

\item{show.ttest}{controls whether results of 2-sample t test is displayed.}
}
\value{
None
}
\description{
Demonstration on judging the difference between two independent samples,
as presented in class.  As shown in class, one part of the demonstration
asks you to judge the difference between the means for two treatment groups.
The second part of the demonstration illustrates the results we obtain
when the data are randomly shuffled over the two groups.
}
\examples{
demo_nhst(demo.type=1) #shows actual observations for 2 groups
demo_nhst(demo.type=2) #shows simulation of random shuffling

}
