% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_permutations.r
\name{demo_permutations}
\alias{demo_permutations}
\title{Demonstration means and variances using permutations}
\usage{
demo_permutations(
  population = c(1:3),
  sample.size = 2,
  pop.mu = NULL,
  pop.sigma2 = NULL,
  show.summary = TRUE,
  show.plots = TRUE,
  show.details = TRUE
)
}
\arguments{
\item{population}{A vector of values to be sampled from}

\item{sample.size}{A number (2-10) for the number of observations to be drawn}

\item{show.summary}{Controls display of summary statistics (TRUE/FALSE)}

\item{show.plots}{Controls whether plots are shown}

\item{show.details}{Controls whether permutations are shown}
}
\value{
ratio of estimated variance/true variance
}
\description{
Demonstration means and variances using permutations
}
\examples{
\dontrun{
demo_permutations(population=c(1:3),sample.size=2)
}
}
