% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.tdist.r
\name{display.tfunction}
\alias{display.tfunction}
\title{Plot a t distribution for dfs 1-100.
Compares t distribution to a normal distribution and shows
the 'excess' area in the tails of the selected t distribution
relative to the standard normal distribution.}
\usage{
display.tfunction(n = 2, plot.normal = TRUE, add.quantile = TRUE)
}
\arguments{
\item{n}{An integer for sample size.}

\item{plot.normal}{Logical.  Sets whether normal is superimposed.}

\item{add.quantile}{Logical.  Sets whether excess area in tails is displayed.}
}
\value{
None
}
\description{
Plot a t distribution for dfs 1-100.
Compares t distribution to a normal distribution and shows
the 'excess' area in the tails of the selected t distribution
relative to the standard normal distribution.
}
\examples{
display.tfunction()

}
