% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.anxiety.r
\name{eda.anxiety}
\alias{eda.anxiety}
\title{Perform two-way anova and related methods using simulated anxiety data.
Typically called by run_anxiety()}
\usage{
eda.anxiety(x = NULL, show.hist = TRUE, show.summary = FALSE,
  show.anova = FALSE, plot.main.gender = FALSE,
  plot.main.condition = FALSE, plot.interaction = FALSE,
  show.tukey = FALSE)
}
\arguments{
\item{x}{A data frame consisting of gender, condition, and anxiety (DV)}

\item{show.hist}{Controls display of initializing histogram}

\item{show.summary}{Controls display of summary statistics}

\item{show.anova}{Controls display of 2-way ANOVA}

\item{plot.main.gender}{Controls display of gender main effect}

\item{plot.main.condition}{Controls display of main effect for treatment}

\item{plot.interaction}{Controls display of interaction}

\item{show.tukey}{Controls display of Tukey HSD for pairwise comparisons}
}
\value{
none
}
\description{
Perform two-way anova and related methods using simulated anxiety data.
Typically called by run_anxiety()
}
\examples{
\dontrun{
eda.anxiety()
}
}
