% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.whr.r
\name{eda.whr}
\alias{eda.whr}
\title{Perform one-way anova and related methods using WHR_DATA.
Typically called by run_whr()}
\usage{
eda.whr(dv, iv, f1.name, f2.name, add.normal = FALSE,
  add.qnorm = FALSE, show.aov = FALSE, show.tukey = FALSE,
  show.barplot = FALSE, conf.level = 0.95, show.boxplots = FALSE,
  n.int = 10, adj.barlimits = FALSE, show.stats = FALSE)
}
\arguments{
\item{dv}{A continuous quantitative dependent variable}

\item{iv}{A categorical variable using as the independent variable}

\item{f1.name}{The name of the dv}

\item{f2.name}{The name of the iv}

\item{add.normal}{Logical. Controls display of normal curve on histogram.}

\item{add.qnorm}{Logical. Controls display of normal quantile plot}

\item{show.aov}{Logical. controls display of ANOVA model}

\item{show.tukey}{Logical. controls display of Tukeys HSD for pairwise comparisons.}

\item{show.barplot}{Controls barplot + CIs of means}

\item{conf.level}{Confidence level used for barplot + CIs}

\item{show.boxplots}{Controls display of boxplots}

\item{n.int}{number of intervals used in histogram}

\item{adj.barlimits}{Controls limits for y axis in barplot}

\item{show.stats}{Controls display of descriptive statistics}
}
\value{
none
}
\description{
Perform one-way anova and related methods using WHR_DATA.
Typically called by run_whr()
}
\examples{
\dontrun{
eda.whr()
}
}
