% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.quantiles.r
\name{get.quantiles}
\alias{get.quantiles}
\title{Get quantiles needed for power analysis}
\usage{
get.quantiles(mu = 60, sd = 1, alpha = 0.05, alt.hyp = "mu1<mu0",
  xlimits = c(54, 64))
}
\arguments{
\item{mu}{A value for the mean of the distribution}

\item{sd}{A positive value for the standard deviation of the distribution}

\item{alpha}{A value between .005 and .1 for the Type I error rate}

\item{alt.hyp}{One of: "mu1<>mu0", "mu1<mu0", or "mu1>mu0"}

\item{xlimits}{Values for the range of the x-axis}
}
\value{
qs The 4 quantiles needed to obtain areas for Ho and Ha
}
\description{
Get quantiles needed for power analysis
}
\examples{
 get.quantiles()

}
