\name{Global.Temps}
\alias{Global.Temps}
\docType{data}
\title{
Global.Temps
}
\description{
Global temperatures, by decades, from 1880-2020, in degrees Centrigrade.
}
\usage{data("Global.Temps")}
\format{
  A data frame with 14 observations on the following 3 variables.
  \describe{
    \item{\code{decade}}{a numeric vector}
    \item{\code{temp.change}}{a numeric vector}
    \item{\code{temp.C}}{a numeric vector}
  }
}
\details{
https://www.climate.gov/news-features/understanding-climate/climate-change-global-temperature
}
\source{
https://www.climate.gov/news-features/understanding-climate/climate-change-global-temperature
}
\references{
https://www.climate.gov/news-features/understanding-climate/climate-change-global-temperature
}
\examples{
data(Global.Temps)
## maybe str(Global.Temps) ; plot(Global.Temps) ...
}
\keyword{datasets}
