% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histnormal.r
\name{histnormal}
\alias{histnormal}
\title{Plot a density histogram with a normal distribution superimposed}
\usage{
histnormal(
  x = rnorm(2000),
  mu = NULL,
  sigma = NULL,
  plot.points = FALSE,
  plot.quantile = FALSE,
  plot.curve = FALSE,
  ran.seed = 4
)
}
\arguments{
\item{x}{#The data to be plotted}

\item{mu}{#The assumed true value for the population mean}

\item{sigma}{#The assumed true value for the population sd}

\item{plot.points}{#Should heights for select points be plotted?}

\item{plot.quantile}{#Should a normal quantile plot be included?}

\item{plot.curve}{#Should normal function be superimposed?}

\item{ran.seed}{#Seed value to allow replication of specific sample}
}
\value{
A data frame consisting of selected quantiles and their heights
}
\description{
Plot a density histogram with a normal distribution superimposed
}
\examples{
histnormal()
histnormal(sunspots)
histnormal(sunspots,plot.points=TRUE,plot.quantile=TRUE)

}
