% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.sample.t.r
\name{one.sample.t}
\alias{one.sample.t}
\title{One-sample Students t test, with display of t function and
areas beyond t observed.}
\usage{
one.sample.t(x = NULL, sample.mean = NULL, sample.sd = NULL,
  n = NULL, mu = 0, test.type = "2-sided")
}
\arguments{
\item{x}{Data to be used}

\item{sample.mean}{A value to be supplied if original observations are not.}

\item{sample.sd}{A value to be supplied if original observations are not.}

\item{n}{The number of observations. (Use when dat are not supplied.)}

\item{mu}{A value for the expected mean under the null hypothesis}

\item{test.type}{2-sided or 1-sided}
}
\value{
Test results
}
\description{
One-sample Students t test, with display of t function and
areas beyond t observed.
}
\examples{
one.sample.t(x=rnorm(10),mu=0)

}
