% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesample.ttest.r
\name{onesample.ttest}
\alias{onesample.ttest}
\title{One sample t test}
\usage{
onesample.ttest(x = rnorm(10, 0, 1), test.type = "2-sided")
}
\arguments{
\item{x}{Data to be tested}

\item{test.type}{one- or two-sided test}
}
\value{
none
}
\description{
One sample t test
}
\examples{
 onesample.ttest(x=rnorm(10),test.type="2-sided")

}
