% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.plot1.r
\name{power.plot1}
\alias{power.plot1}
\title{Plot the null or alternative sampling distribution for a power analysis}
\usage{
power.plot1(mu = 60, sd = 1, xlimits = c(55, 65), plot.type = "Ho",
  alpha = 0.05, xcors = c(56, 59, 61, 65), alt.hyp = "mu1<mu0",
  shade.colors = c("darkgoldenrod1", "cornsilk", "darkgoldenrod1"),
  superimpose = FALSE, show.values = FALSE, show.means = FALSE)
}
\arguments{
\item{mu}{A value for the mean of the distribution}

\item{sd}{A positive value for the standard deviation of the distribution}

\item{xlimits}{Values for the range of the x-axis}

\item{plot.type}{One of: "Ho" or "Ha" for the type of distribution to plot}

\item{alpha}{A value between .005 and .1 for the Type I error rate}

\item{xcors}{The 4 quantiles that delimit the areas of interest}

\item{alt.hyp}{One of: "mu1<>mu0", "mu1<mu0", or "mu1>mu0"}

\item{shade.colors}{colors for the regions of interest}

\item{superimpose}{Whether to create new plot for Ha or superimpose over Ho}

\item{show.values}{Select whether Type I, II, and Power are displayed in plot.}

\item{show.means}{Whether to plot means for Ho and Ha}
}
\value{
pow Probability of correct rejection (For plot.type="Ha")
}
\description{
Plot the null or alternative sampling distribution for a power analysis
}
\examples{
 power.plot1()

}
