% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.r
\name{power1}
\alias{power1}
\title{Perform a power analysis, with plots for null and alternative distributions}
\usage{
power1(mu0 = 50, mu1 = 50, sd = 10, n = 100, alpha = 0.05,
  alt.hyp = "mu1<>mu0")
}
\arguments{
\item{mu0}{A value for the mean under the null hypothesis}

\item{mu1}{A value for the mean under the alternative hypothesis}

\item{sd}{A positive value for the standard deviation for the population (null)}

\item{n}{A value for the sample size}

\item{alpha}{A value between .005 and .1 for the Type I error rate}

\item{alt.hyp}{One of: "mu1<>mu0", "mu1<mu0", or "mu1>mu0"}
}
\value{
res Summary, included Type I error (alpha), Type II error (beta), and Power
}
\description{
Perform a power analysis, with plots for null and alternative distributions
}
\examples{
 power1()

}
