% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.anxiety.r
\name{run.anxiety}
\alias{run.anxiety}
\title{manipulate two-way anova and related methods using eda.anxiety}
\usage{
run.anxiety()
}
\value{
none
}
\description{
manipulate two-way anova and related methods using eda.anxiety
}
\examples{
\dontrun{
run.anxiety()
}
}
