% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.binomials1.r
\name{run.binomials1}
\alias{run.binomials1}
\title{Demonstrate binomials using manipulate controls that allow
you to set number of flips, probability of success, whether
the plot is counts or probabilities, and whether a Gaussian
distribution is superimposed.  Also calculates areas to the
left and right of critical values for counts or probabilities.}
\usage{
run.binomials1()
}
\description{
Demonstrate binomials using manipulate controls that allow
you to set number of flips, probability of success, whether
the plot is counts or probabilities, and whether a Gaussian
distribution is superimposed.  Also calculates areas to the
left and right of critical values for counts or probabilities.
}
\examples{
\dontrun{
binomials1()
}
}
