% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.histnormal.r
\name{run.histnormal}
\alias{run.histnormal}
\title{Manipulate histnormal function}
\usage{
run.histnormal(x = rnorm(1000))
}
\arguments{
\item{x}{Data to be plotted}
}
\value{
None
}
\description{
Manipulate histnormal function
}
\examples{
\dontrun{
histnormal()
}

}
