% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.onesample.ztest.r
\name{run.onesample.ztest}
\alias{run.onesample.ztest}
\title{Manipulate one-sample z test}
\usage{
run.onesample.ztest(mu = 100, sigma = 15, n = 10,
  observed.mean = 110)
}
\arguments{
\item{mu}{A value for assumed true mean of population}

\item{sigma}{A value for assume true sigma of population}

\item{n}{An integer for sample size}

\item{observed.mean}{A value for the observed sample size}
}
\value{
None
}
\description{
Manipulate one-sample z test
}
\examples{
\dontrun{
run.onesample.ztest()
}

}
