% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.sampling.dist1.r
\name{run.sampling.dist1}
\alias{run.sampling.dist1}
\title{Demonstration of sampling distribution of the mean}
\usage{
run.sampling.dist1()
}
\value{
None
}
\description{
Demonstration of sampling distribution of the mean
}
\examples{
\dontrun{
run.sampling.dist1()
}

}
