% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tdist.r
\name{run.tdist}
\alias{run.tdist}
\title{Manipulate t distribution.
Displays a t distribtution for dfs 1-100, compares t distribtution
to normal distribution if desired, and shows 'excess' area in the tails
of a t distribution relative to a standard normal distribution.}
\usage{
run.tdist()
}
\value{
None
}
\description{
Manipulate t distribution.
Displays a t distribtution for dfs 1-100, compares t distribtution
to normal distribution if desired, and shows 'excess' area in the tails
of a t distribution relative to a standard normal distribution.
}
\examples{
\dontrun{
run.tdist()
}

}
