% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.whew.r
\name{run.whew}
\alias{run.whew}
\title{manipulate whew1 for regression exercise using Health, Education, and GDP data}
\usage{
run.whew()
}
\value{
none
}
\description{
manipulate whew1 for regression exercise using Health, Education, and GDP data
}
\examples{
\dontrun{
run.whew()
}
}
