% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.whr.r
\name{run.whr}
\alias{run.whr}
\title{manipulate one-way anova and related methods using WHR_DATA}
\usage{
run.whr()
}
\value{
none
}
\description{
manipulate one-way anova and related methods using WHR_DATA
}
\examples{
\dontrun{
run.whr()
}
}
