% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.dist1.r
\name{samp.dist1}
\alias{samp.dist1}
\title{Demonstration of sampling distribution of the mean}
\usage{
samp.dist1(
  population = NULL,
  ssize = 5,
  pop.type = "Uniform[1-10]",
  add.normal = TRUE,
  custom.data = NULL,
  show.pop = TRUE,
  show.flex = TRUE,
  show.fixed = TRUE,
  show.qnorm = TRUE
)
}
\arguments{
\item{population}{An optional numeric vector to be treated as the population}

\item{ssize}{An integer for the size of the sample}

\item{pop.type}{Controls the type of population from which samples are drawn}

\item{add.normal}{#controls whether normal distribution is superimposed}

\item{custom.data}{User supplied vector of data to be used}

\item{show.pop}{display population}

\item{show.flex}{show sampling distribution with flexible x axis}

\item{show.fixed}{show sampling distribution with fixed x axis}

\item{show.qnorm}{show normal quantile plot}
}
\value{
None
}
\description{
Demonstration of sampling distribution of the mean
}
\examples{
samp.dist1()

}
