% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpop.r
\name{sdpop}
\alias{sdpop}
\title{Return sigma for assumed population}
\usage{
sdpop(x)
}
\arguments{
\item{x}{a vector consisting of the elements of the population}
}
\value{
sigma for the population
}
\description{
Return sigma for assumed population
}
\examples{
sdpop(x=c(1:6))

}
