% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.ci.r
\name{show.ci}
\alias{show.ci}
\title{Plot Confidence Intervals for CLPS0900}
\usage{
show.ci(nsamp = 100, ssize = 2, mu = 0, sigma = 1,
  conf.level = 0.68, do.pause = "None", ci.method = "Sigma known",
  show.data = FALSE, xaxis.type = "Fixed", go.button = NULL)
}
\arguments{
\item{nsamp}{Integer for number of samples to display}

\item{ssize}{The size of each sample}

\item{mu}{Population mean}

\item{sigma}{Population sigma}

\item{conf.level}{Confidence level}

\item{do.pause}{Whether to pause between each CI}

\item{ci.method}{Whether sigma is known or unknown}

\item{show.data}{Whether sample data are displayed}

\item{xaxis.type}{Whether x axis if fixed or variable}

\item{go.button}{Used for manipulate only}
}
\value{
None
}
\description{
Plot Confidence Intervals for CLPS0900
}
\examples{
show.ci()

}
