\name{WHEW_DATA}
\alias{WHEW_DATA}
\docType{data}
\title{
WHEW_DATA
}
\description{
Data from World Happiness Report 2018 and UN GDP data. Formatted for use by students in
CLPS0900 (Quantitative Methods in the Psychological Sciences).
}
\usage{data("WHEW_DATA")}
\format{
A data frame consisting of 106 rows (countries) and 5 variables

}
\details{
 See: http://worldhappiness.report/ed/2019/
}
\source{
http://worldhappiness.report/ed/2019/
}
\references{
http://worldhappiness.report/ed/2019/
}
\examples{
data(WHEW_DATA)
## maybe mplot(WHEW_DATA)
}
\keyword{datasets}
