\name{WHR_DATA}
\alias{WHR_DATA}
\docType{data}
\title{
WHR_DATA
}
\description{
WHR data from World Happiness Report 2018. Formatted for use by students in
CLPS0900 (Quantitative Methods in the Psychological Science).
}
\usage{data("WHR_DATA")}
\format{
A data frame consisting of 192 rows (countries) and 6 variables

}
\details{
More info about data set goes here...
}
\source{
http://worldhappiness.report/ed/2019/
}
\references{
http://worldhappiness.report/ed/2019/
}
\examples{
data(WHR_DATA)
## maybe mplot(WHR_DATA)
}
\keyword{datasets}
