% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPS0900.R
\docType{package}
\name{CLPS0900R}
\alias{CLPS0900R}
\title{CLPS0900R: Utilities for Quantitative Methods course.}
\description{
The package includes demonstrations in R that are presented in class,
tools that are used in weekly problem sets, and associated data sets.
}
\section{CLPS0900R functions}{

run.whr() A suite of ANOVA and other analyses using World Happiness Report data

demo_hist()  A simple histogram

demo_nhst()  A simulation of 2-sample comparisons

sdpop()  Returns sigma for an assumed population of values

binomials1()  Shows binomial distribution

run.binomials1()  Run binomial demonstration with manipulate wrapper

samp.dist1()  Show a sampling distribution of the mean

run.sampling.dist1()  Run sampling distribution with manipulate wrapper

histnormal()  Create density histogram with normal superimposed

run.histnormal() Manipulate histnormal() function

onesample.ztest() NHST of one sample mean using known mu and sigma

run.onesample.ztest() Manipulate onesample.ztest

one.sample.t() Another version of one sample Students t test, allowing different inputs

run.power() Perform a power analysis

run.tdist() Manipulate a t distribution and compare to normal
display.tfunction() Plot Students t and compare to standard normal
}

