% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPS0900.R
\docType{package}
\name{CLPS0900R}
\alias{CLPS0900R}
\alias{CLPS0900R-package}
\title{CLPS0900R: Utilities for Quantitative Methods course.}
\description{
The package includes demonstrations in R that are presented in class,
tools that are used in weekly problem sets, and associated data sets.
}
\details{
\describe{
  \item{demo_hist()}{A simple histogram}
  \item{demo_nhst()}{A simulation of 2-sample comparisons}
  \item{sdpop(x)}{Returns sigma for an assumed population of values}
  \item{binomials1()}{Shows binomial distribution}
}
}
