% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPS0900.R
\docType{package}
\name{CLPS0900R}
\alias{CLPS0900R}
\alias{CLPS0900R-package}
\title{CLPS0900R: Utilities for Quantitative Methods course.}
\description{
The package includes demonstrations in R that are presented in class,
and tools that are used in weekly problem sets.
}
