% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histnormal.r
\name{histnormal}
\alias{histnormal}
\title{Plot a density histogram with a normal distribution superimposed}
\usage{
histnormal(x = rnorm(1000), mu = NULL, sigma = NULL,
  plot.points = FALSE, plot.quantile = FALSE)
}
\arguments{
\item{x}{#The data to be plotted}

\item{mu}{#The assumed true value for the population mean}

\item{sigma}{#The assumed true value for the population sd}

\item{plot.points}{#Should heights for select points be plotted?}

\item{plot.quantile}{#Should a normal quantile plot be included?}
}
\value{
A data frame consisting of selected quantiles and their heights
}
\description{
Plot a density histogram with a normal distribution superimposed
}
\examples{
histnormal()
histnormal(sunspots)
histnormal(sunspots,plot.points=TRUE,plot.quantile=TRUE)

}
