% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesample.ztest.r
\name{onesample.ztest}
\alias{onesample.ztest}
\title{One sample NHST with known mu and sigma}
\usage{
onesample.ztest(mu = 100, sigma = 15, n = 10, observed.mean = 110,
  test.type = "2-sided")
}
\arguments{
\item{mu}{A value for assumed population true mean}

\item{sigma}{A value for assumed population true sigma}

\item{n}{An integer for sample size}

\item{observed.mean}{A value for the observed sample mean}

\item{test.type}{one- or two-sided test}
}
\value{
none
}
\description{
One sample NHST with known mu and sigma
}
\examples{
 onesample.ztest(mu=100,sigma=15,n=10,observed.mean=110,test.type="2-sided")

}
