% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesample.ztest.r
\name{onesample.ztest}
\alias{onesample.ztest}
\title{One sample NHST with known mu and sigma}
\usage{
onesample.ztest(
  observed.mean = 110,
  mu = 100,
  sigma = 15,
  n = 10,
  test.type = "2-sided",
  conf.level = 0.95
)
}
\arguments{
\item{observed.mean}{A value for the observed sample mean if data not provided.}

\item{mu}{A value for assumed population true mean}

\item{sigma}{A value for assumed population true sigma}

\item{n}{An integer for sample size (if data not provided).}

\item{test.type}{one- or two-sided test}

\item{conf.level}{A value for confidence level for interval (.68-.95)}
}
\value{
none
}
\description{
One sample NHST with known mu and sigma
}
\examples{
 onesample.ztest(observed.mean=110,n=4,mu=100,sigma=15)

}
