% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.power.r
\name{run.power}
\alias{run.power}
\title{Manipulate plot the null or alternative sampling distribution for a power analysis}
\usage{
run.power(mu0 = 60, mu1 = 57, sd = 10, n = 100)
}
\arguments{
\item{mu0}{A value for the mean under the null hypothesis}

\item{mu1}{A value for the mean under the alternative hypothesis}

\item{sd}{A positive value for the standard deviation for the population (null)}

\item{n}{A value for the sample size}
}
\value{
res Summary, included Type I error (alpha), Type II error (beta), and Power
}
\description{
Manipulate plot the null or alternative sampling distribution for a power analysis
}
\examples{
\dontrun{
run.power()
}
}
