% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.dist1.r
\name{samp.dist1}
\alias{samp.dist1}
\title{Demonstration of sampling distribution of the mean}
\usage{
samp.dist1(ssize = 5, pop.type = "Uniform[1-10]", add.normal = TRUE)
}
\arguments{
\item{ssize}{An integer for the size of the sample}

\item{pop.type}{Controls the type of population from which samples are drawn}

\item{add.normal}{#controls whether normal distribution is superimposed}
}
\value{
None
}
\description{
Demonstration of sampling distribution of the mean
}
\examples{
samp.dist1()

}
